Continue factoring out offer input
